--- JourneyHelper-Web/WebRoot/WEB-INF/classes/.gitignore
@@ -1,1 +1,0 @@
-/com
